Pass Best Bet keywords through "filter_query" method
When loading the Best Bet keywords, pass them through the "filter_query"
method in the "QuickSearch::QueryFilter" concern, so that they will be
matched with a query term passed  through the same method. This takes
care of things like a dash (hyphen) being removed from the query term.

The implementation was somewhat complicated by the fact that the
"filter_query" method is a private method on a Rails concern. This was
handled by creating a helper class in the initializer that delegates to
the concern. It is not clear if this is the best way to accomplish this,
but it does appear to work.

Modified "test/dummy/config/best_bets.yml" to add a new Best Bet entry
with a keyword including a dash for use in
"test/integration/best_bets_initializer_test.rb"

Added "xhr_search" route to "test/dummy/config/routes.rb" so it
would be available as a "*_path" helper in the test.